requirements.txt
adrf==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
amqp==5.3.1
anyio==3.7.1
asgiref==3.10.0
asttokens==3.0.0
async-property==0.2.2
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
billiard==4.2.2
cachetools==6.2.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
decorator==5.2.1
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
executing==2.2.1
frozenlist==1.7.0
gensim==4.3.3
google-auth==2.41.1
google-auth-oauthlib==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
kombu==5.5.4
matplotlib-inline==0.1.7
multidict==6.6.4
neo4j==5.19.0
neomodel==5.3.1
numpy==1.26.4
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.4.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pylast==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2

requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.13.1
setuptools==80.9.0
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spotipy==2.25.1
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
websockets==12.0
wheel==0.45.1
wrapt==1.17.3
yarl==1.20.1
ytmusicapi==1.11.1
